flight-booking/requirements.txt
argon2-cffi==18.3.0
autopep8==1.4.3
bcrypt==3.1.4
celery==4.2.0
cloudinary==1.15.0
coverage==4.5.2
Django==2.1.5
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-middleware==1.3.1
django-environ==0.4.5
django-heroku==0.3.1
django-redis-cache==1.8.1
django-nose==1.4.6
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
isort==4.3.4
mysql-connector-python==8.0.14
mysqlclient==1.4.1
nose==1.3.7
pep8==1.7.1
pika==0.9.14
pillow==5.4.1
psycopg2-binary==2.7.6.1
pylint==2.1.1
pytz==2018.7
redis==3.0.1
requests>=2.13.0
